- saneCheck now check for missing extensions - saneCheck now check for speed deviation and... - saneFix can flag for it
- error code have been changed to uint32_t to accomodate for the flags of ErrorCodes.h


git-svn-id: https://svn.ias.u-psud.fr/sanepic/trunk@1403 339b86bc-bb6a-0410-a225-86439261ec80

+++ saneFrameOrder/src/FrameOrder.cpp
@@ -24,7 +24,7 @@ using namespace std;
 #include "Utilities.h"
 #include "MPIConfiguration.h"
 
-uint16_t writeParallelScheme(string outdir, vector<long> order, struct samples samples_struct)
+uint32_t writeParallelScheme(string outdir, vector<long> order, struct samples samples_struct)
 // Write the Parallelization Scheme for further use.
 {
 
@@ -125,14 +125,14 @@ void  AssignNodeFloatWeight(vector<string> nodeName, map<string, float> map_node
 
 }
 
-uint16_t removeProcName(std::string outdir){
+uint32_t removeProcName(std::string outdir){
 	string filename;
 	filename = outdir+processorName_filename;
 	return remove(filename.c_str());
 
 }
 
-uint16_t readNodeWeight(string & output, string pathIn, map<string, float> & nodeWeight){
+uint32_t readNodeWeight(string & output, string pathIn, map<string, float> & nodeWeight){
   /**
    * read node weight into an associative hash
    */
