+++ saneInv/src/main.cpp
@@ -106,11 +106,11 @@ int main(int argc, char *argv[]) {
 
 	std::vector<int> indexIn; /* bolometer index, used to determine which intput detector corresponds to which output detector*/
 
-	uint16_t mask_saneInv = INI_NOT_FOUND | DATA_INPUT_PATHS_PROBLEM | TMP_PATH_PROBLEM |
+	uint32_t mask_saneInv = INI_NOT_FOUND | DATA_INPUT_PATHS_PROBLEM | TMP_PATH_PROBLEM |
 			BOLOFILE_NOT_FOUND | SANEINV_INPUT_ERROR | FITS_FILELIST_NOT_FOUND; // 0x601b
 
-	uint16_t parsed=0x0000; // parser error status
-	uint16_t compare_to_mask; // parser error status
+	uint32_t parsed=0x0000; // parser error status
+	uint32_t compare_to_mask; // parser error status
 
 	//	if (rank==0){ // root parse ini file and fill the structures. Also print warnings or errors
