+++ sanePS/src/main.cpp
@@ -121,7 +121,7 @@ int main(int argc, char *argv[]){
 	std::ostringstream temp_stream; // used to remove sprintf horror
 
 
-	uint16_t mask_sanePS = INI_NOT_FOUND | DATA_INPUT_PATHS_PROBLEM | OUPUT_PATH_PROBLEM | TMP_PATH_PROBLEM |
+	uint32_t mask_sanePS = INI_NOT_FOUND | DATA_INPUT_PATHS_PROBLEM | OUPUT_PATH_PROBLEM | TMP_PATH_PROBLEM |
 			BOLOFILE_NOT_FOUND | FSAMP_PROBLEM | NCOMP_WRONG_VALUE | ELL_FILE_NOT_FOUND | MIX_FILE_NOT_FOUND |
 			FITS_FILELIST_NOT_FOUND | FCUT_PROBLEM; // 0xdd1f
 
@@ -130,8 +130,8 @@ int main(int argc, char *argv[]){
 
 	// parser variables
 	int indice_argv = 1;
-	uint16_t parsed=0x0000; // parser error status
-	uint16_t compare_to_mask; // parser error status
+	uint32_t parsed=0x0000; // parser error status
+	uint32_t compare_to_mask; // parser error status
 
 	PS_param.restore = 0; //default
